feat(study): 가사 서버연결
