+++ components/study/PlayController.tsx
@@ -184,8 +184,8 @@ const Styled = {
     justify-content: center;
     width: 100%;
     .bar {
-      align-self: center;
       -webkit-appearance: none;
+      align-self: center;
       margin-bottom: 20px;
       border-radius: 10px;
       background-color: #9d9d9d;
@@ -224,7 +224,7 @@ const Styled = {
     align-items: center;
     justify-content: center;
     .back-btn {
-      background: url('assets/icons/backIcon.svg') no-repeat 0 0;
+      background: url('/assets/icons/backIcon.svg') no-repeat 0 0;
       width: 19px;
       height: 23px;
       &:hover {
@@ -238,10 +238,10 @@ const Styled = {
       ${({ isPlay }) =>
         isPlay
           ? css`
-              background: url('assets/icons/pauseIcon.svg') no-repeat 0 0;
+              background: url('/assets/icons/pauseIcon.svg') no-repeat 0 0;
             `
           : css`
-              background: url('assets/icons/playIcon.svg') no-repeat 0 0;
+              background: url('/assets/icons/playIcon.svg') no-repeat 0 0;
             `}
       width: 31px;
       height: 31px;
@@ -251,7 +251,7 @@ const Styled = {
       }
     }
     .forward-btn {
-      background: url('assets/icons/forwardIcon.svg') no-repeat 0 0;
+      background: url('/assets/icons/forwardIcon.svg') no-repeat 0 0;
       width: 19px;
       height: 23px;
       &:hover {
@@ -306,7 +306,7 @@ const Styled = {
       width: 116.01px;
       height: 25px;
       &__btn {
-        background: url('assets/icons/soundIcon.svg') no-repeat 0 0;
+        background: url('/assets/icons/soundIcon.svg') no-repeat 0 0;
         width: 25px;
         height: 25px;
         &:hover {
@@ -341,7 +341,7 @@ const Styled = {
       justify-content: center;
       &__btn {
         margin-bottom: 10px;
-        background: url('assets/icons/replayIcon.svg') no-repeat 0 0;
+        background: url('/assets/icons/replayIcon.svg') no-repeat 0 0;
         width: 20px;
         height: 20px;
         ${({ isLooped }) =>
